fix piggy banks
